PLAT-98958: Restructuring VirtualList and Scroller with Hooks (Phase 2) (#2682)
* Rename useScroll as useScrollable and useChildPropsDecorator as useScroll

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Remove useForceUpdate.js and define useForceUpdate in Scrollable.js

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Delete useForceUpdate.js

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Delete utilEventResizeWindow.js

* Reduce the diff from develop branch

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Reduce the diff from develop branch

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Reduce the diff from develop branch

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Reduce the diff from develop branch

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Reduce the diff from develop branch

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Reduce num of files

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Reduce diff

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Fix adapter and containerRef

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Fix adapter and containerRef

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Fix ref

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Cleanup

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Cleanup

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Cleanup

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Rename scrollableContainerRef as uiScrollableContainerRef,

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Cleanup

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Cleanup

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Rename variables as mutableRef

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Fix lint

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Fix typo and errors

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Fix adapter and ref

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Cleanup

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Rename

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Rename

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Fix

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Fix

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Rename

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Fix

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Cleanup

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Apply PLAT-95353

* Fix

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Fix lint

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Fix uiScrollAdapter sync

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Cleanup

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Cleanup

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Fix lint

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Fix cbScrollTo

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Fix lint

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Fix

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)
