+++ packages/moonstone/Scrollable/Scrollable.js
@@ -6,176 +6,1320 @@
  * @private
  */
 
+import {forward} from '@enact/core/handle';
+import platform from '@enact/core/platform';
+import Spotlight from '@enact/spotlight';
+import {spottableClass} from '@enact/spotlight/Spottable';
+import {getTargetByDirectionFromPosition} from '@enact/spotlight/src/target';
+import {getRect, intersects} from '@enact/spotlight/src/utils';
+import {useScrollBase} from '@enact/ui/Scrollable';
+import {useChildAdapter as useUiChildAdapter} from '@enact/ui/Scrollable/useChild';
+import {utilDecorateChildProps} from '@enact/ui/Scrollable';
+import utilDOM from '@enact/ui/Scrollable/utilDOM';
+import utilEvent from '@enact/ui/Scrollable/utilEvent';
 import PropTypes from 'prop-types';
+import React, {Component, useContext, useRef} from 'react';
 
-import {dataIndexAttribute} from './useSpottableScrollable';
+import $L from '../internal/$L';
+import {SharedState} from '../internal/SharedStateDecorator';
 
-// ScrollableBase is now only used in storybook.
-const ScrollableBase = {};
+import {useChildAdapter} from './useChild';
+import {
+	useEventFocus, useEventKey, useEventMonitor, useEventMouse,
+	useEventTouch, useEventVoice, useEventWheel
+} from './useEvent';
+import useOverscrollEffect from './useOverscrollEffect';
+import useScrollbar from './useScrollbar';
+import {useSpotlightConfig, useSpotlightRestore} from './useSpotlight';
 
-ScrollableBase.displayName = 'Scrollable';
+import overscrollCss from './OverscrollEffect.module.less';
 
-ScrollableBase.propTypes = /** @lends moonstone/Scrollable.Scrollable.prototype */ {
-	/**
-	 * Render function.
-	 *
-	 * @type {Function}
-	 * @required
-	 * @private
-	 */
-	childRenderer: PropTypes.func.isRequired,
+const
+	reverseDirections = {
+		down: 'up',
+		up: 'down'
+	};
 
-	/**
-	 * This is set to `true` by SpotlightContainerDecorator
-	 *
-	 * @type {Boolean}
-	 * @private
-	 */
-	'data-spotlight-container': PropTypes.bool,
+/**
+ * The name of a custom attribute which indicates the index of an item in
+ * [VirtualList]{@link moonstone/VirtualList.VirtualList} or
+ * [VirtualGridList]{@link moonstone/VirtualList.VirtualGridList}.
+ *
+ * @constant dataIndexAttribute
+ * @memberof moonstone/Scrollable
+ * @type {String}
+ * @private
+ */
+const dataIndexAttribute = 'data-index';
 
-	/**
-	 * `false` if the content of the list or the scroller could get focus
-	 *
-	 * @type {Boolean}
-	 * @default false
-	 * @private
-	 */
-	'data-spotlight-container-disabled': PropTypes.bool,
-
-	/**
-	 * This is passed onto the wrapped component to allow
-	 * it to customize the spotlight container for its use case.
-	 *
-	 * @type {String}
-	 * @private
-	 */
-	'data-spotlight-id': PropTypes.string,
-
-	/**
-	 * Direction of the list or the scroller.
-	 * `'both'` could be only used for[Scroller]{@link moonstone/Scroller.Scroller}.
-	 *
-	 * Valid values are:
-	 * * `'both'`,
-	 * * `'horizontal'`, and
-	 * * `'vertical'`.
-	 *
-	 * @type {String}
-	 * @private
-	 */
-	direction: PropTypes.oneOf(['both', 'horizontal', 'vertical']),
-
-	/**
-	 * Allows 5-way navigation to the scrollbar controls. By default, 5-way will
-	 * not move focus to the scrollbar controls.
-	 *
-	 * @type {Boolean}
-	 * @default false
-	 * @public
-	 */
-	focusableScrollbar: PropTypes.bool,
-
-	/**
-	 * A unique identifier for the scrollable component.
-	 *
-	 * When specified and when the scrollable is within a SharedStateDecorator, the scroll
-	 * position will be shared and restored on mount if the component is destroyed and
-	 * recreated.
-	 *
-	 * @type {String}
-	 * @public
-	 */
-	id: PropTypes.string,
-
-	/**
-	 * Specifies overscroll effects shows on which type of inputs.
-	 *
-	 * @type {Object}
-	 * @default {
-	 *	arrowKey: false,
-	 *	drag: false,
-	 *	pageKey: false,
-	 *	scrollbarButton: false,
-	 *	wheel: true
-	 * }
-	 * @private
-	 */
-	overscrollEffectOn: PropTypes.shape({
-		arrowKey: PropTypes.bool,
-		drag: PropTypes.bool,
-		pageKey: PropTypes.bool,
-		scrollbarButton: PropTypes.bool,
-		wheel: PropTypes.bool
-	}),
-
-	/**
-	 * Specifies preventing keydown events from bubbling up to applications.
-	 * Valid values are `'none'`, and `'programmatic'`.
-	 *
-	 * When it is `'none'`, every keydown event is bubbled.
-	 * When it is `'programmatic'`, an event bubbling is not allowed for a keydown input
-	 * which invokes programmatic spotlight moving.
-	 *
-	 * @type {String}
-	 * @default 'none'
-	 * @private
-	 */
-	preventBubblingOnKeyDown: PropTypes.oneOf(['none', 'programmatic']),
-
-	/**
-	 * Sets the hint string read when focusing the next button in the vertical scroll bar.
-	 *
-	 * @type {String}
-	 * @default $L('scroll down')
-	 * @public
-	 */
-	scrollDownAriaLabel: PropTypes.string,
-
-	/**
-	 * Sets the hint string read when focusing the previous button in the horizontal scroll bar.
-	 *
-	 * @type {String}
-	 * @default $L('scroll left')
-	 * @public
-	 */
-	scrollLeftAriaLabel: PropTypes.string,
-
-	/**
-	 * Sets the hint string read when focusing the next button in the horizontal scroll bar.
-	 *
-	 * @type {String}
-	 * @default $L('scroll right')
-	 * @public
-	 */
-	scrollRightAriaLabel: PropTypes.string,
-
-	/**
-	 * Sets the hint string read when focusing the previous button in the vertical scroll bar.
-	 *
-	 * @type {String}
-	 * @default $L('scroll up')
-	 * @public
-	 */
-	scrollUpAriaLabel: PropTypes.string,
-
-	/**
-	 * TBD
-	 */
-	type: PropTypes.string
+const isIntersecting = (elem, container) => elem && intersects(getRect(container), getRect(elem));
+const getIntersectingElement = (elem, container) => isIntersecting(elem, container) && elem;
+const getTargetInViewByDirectionFromPosition = (direction, position, container) => {
+	const target = getTargetByDirectionFromPosition(direction, position, Spotlight.getActiveContainer());
+	return getIntersectingElement(target, container);
 };
 
-ScrollableBase.defaultProps = {
-	'data-spotlight-container-disabled': false,
-	focusableScrollbar: false,
-	overscrollEffectOn: {
-		arrowKey: false,
-		drag: false,
-		pageKey: false,
-		scrollbarButton: false,
-		wheel: true
-	},
-	preventBubblingOnKeyDown: 'none',
-	type: 'JS'
+/**
+ * A Moonstone-styled component that provides horizontal and vertical scrollbars.
+ *
+ * @class ScrollableBase
+ * @memberof moonstone/Scrollable
+ * @extends ui/Scrollable.ScrollableBase
+ * @ui
+ * @public
+ */
+class ScrollableBase extends Component { // ScrollableBase is now only used in storybook.
+	static displayName = 'Scrollable'
+
+	static propTypes = /** @lends moonstone/Scrollable.Scrollable.prototype */ {
+		/**
+		 * Render function.
+		 *
+		 * @type {Function}
+		 * @required
+		 * @private
+		 */
+		childRenderer: PropTypes.func.isRequired,
+
+		/**
+		 * This is set to `true` by SpotlightContainerDecorator
+		 *
+		 * @type {Boolean}
+		 * @private
+		 */
+		'data-spotlight-container': PropTypes.bool,
+
+		/**
+		 * `false` if the content of the list or the scroller could get focus
+		 *
+		 * @type {Boolean}
+		 * @default false
+		 * @private
+		 */
+		'data-spotlight-container-disabled': PropTypes.bool,
+
+		/**
+		 * This is passed onto the wrapped component to allow
+		 * it to customize the spotlight container for its use case.
+		 *
+		 * @type {String}
+		 * @private
+		 */
+		'data-spotlight-id': PropTypes.string,
+
+		/**
+		 * Direction of the list or the scroller.
+		 * `'both'` could be only used for[Scroller]{@link moonstone/Scroller.Scroller}.
+		 *
+		 * Valid values are:
+		 * * `'both'`,
+		 * * `'horizontal'`, and
+		 * * `'vertical'`.
+		 *
+		 * @type {String}
+		 * @private
+		 */
+		direction: PropTypes.oneOf(['both', 'horizontal', 'vertical']),
+
+		/**
+		 * Allows 5-way navigation to the scrollbar controls. By default, 5-way will
+		 * not move focus to the scrollbar controls.
+		 *
+		 * @type {Boolean}
+		 * @default false
+		 * @public
+		 */
+		focusableScrollbar: PropTypes.bool,
+
+		/**
+		 * A unique identifier for the scrollable component.
+		 *
+		 * When specified and when the scrollable is within a SharedStateDecorator, the scroll
+		 * position will be shared and restored on mount if the component is destroyed and
+		 * recreated.
+		 *
+		 * @type {String}
+		 * @public
+		 */
+		id: PropTypes.string,
+
+		/**
+		 * Specifies overscroll effects shows on which type of inputs.
+		 *
+		 * @type {Object}
+		 * @default {
+		 *	arrowKey: false,
+		 *	drag: false,
+		 *	pageKey: false,
+		 *	scrollbarButton: false,
+		 *	wheel: true
+		 * }
+		 * @private
+		 */
+		overscrollEffectOn: PropTypes.shape({
+			arrowKey: PropTypes.bool,
+			drag: PropTypes.bool,
+			pageKey: PropTypes.bool,
+			scrollbarButton: PropTypes.bool,
+			wheel: PropTypes.bool
+		}),
+
+		/**
+		 * Specifies preventing keydown events from bubbling up to applications.
+		 * Valid values are `'none'`, and `'programmatic'`.
+		 *
+		 * When it is `'none'`, every keydown event is bubbled.
+		 * When it is `'programmatic'`, an event bubbling is not allowed for a keydown input
+		 * which invokes programmatic spotlight moving.
+		 *
+		 * @type {String}
+		 * @default 'none'
+		 * @private
+		 */
+		preventBubblingOnKeyDown: PropTypes.oneOf(['none', 'programmatic']),
+
+		/**
+		 * Sets the hint string read when focusing the next button in the vertical scroll bar.
+		 *
+		 * @type {String}
+		 * @default $L('scroll down')
+		 * @public
+		 */
+		scrollDownAriaLabel: PropTypes.string,
+
+		/**
+		 * Sets the hint string read when focusing the previous button in the horizontal scroll bar.
+		 *
+		 * @type {String}
+		 * @default $L('scroll left')
+		 * @public
+		 */
+		scrollLeftAriaLabel: PropTypes.string,
+
+		/**
+		 * Sets the hint string read when focusing the next button in the horizontal scroll bar.
+		 *
+		 * @type {String}
+		 * @default $L('scroll right')
+		 * @public
+		 */
+		scrollRightAriaLabel: PropTypes.string,
+
+		/**
+		 * Sets the hint string read when focusing the previous button in the vertical scroll bar.
+		 *
+		 * @type {String}
+		 * @default $L('scroll up')
+		 * @public
+		 */
+		scrollUpAriaLabel: PropTypes.string,
+
+		/*
+		 * TBD
+		 */
+		type: PropTypes.string
+	}
+
+	static defaultProps = {
+		'data-spotlight-container-disabled': false,
+		focusableScrollbar: false,
+		overscrollEffectOn: {
+			arrowKey: false,
+			drag: false,
+			pageKey: false,
+			scrollbarButton: false,
+			wheel: true
+		},
+		preventBubblingOnKeyDown: 'none',
+		type: 'JS'
+	}
+}
+
+/* eslint-disable indent */
+/*
+// Move to useSpotlightRestore
+
+	// Only intended to be used within componentDidMount, this method will fetch the last stored
+	// scroll position from SharedState and scroll (without animation) to that position
+	function restoreScrollPosition () {
+		const {id} = props;
+		if (id && context && context.get) {
+			const scrollPosition = context.get(`${id}.scrollPosition`);
+
+			if (scrollPosition) {
+				uiScrollAdapter.current.scrollTo({
+					position: scrollPosition,
+					animate: false
+				});
+			}
+		}
+	}
+
+// Move to useScrollbar
+
+	function isScrollButtonFocused () {
+		return (
+			horizontalScrollbarRef.current && horizontalScrollbarRef.current.isOneOfScrollButtonsFocused() ||
+			verticalScrollbarRef.current && verticalScrollbarRef.current.isOneOfScrollButtonsFocused()
+		);
+	}
+
+// Move to useEventMouse
+
+	function handleFlick ({direction}) {
+		const
+			{canScrollHorizontally, canScrollVertically} = uiScrollAdapter.current,
+			bounds = uiScrollAdapter.current.getScrollBounds(),
+			focusedItem = Spotlight.getCurrent();
+
+		if (focusedItem) {
+			focusedItem.blur();
+		}
+
+		if ((
+			direction === 'vertical' && canScrollVertically(bounds) ||
+			direction === 'horizontal' && canScrollHorizontally(bounds)
+		) && !props['data-spotlight-container-disabled']) {
+			childAdapter.current.setContainerDisabled(true);
+		}
+	}
+
+	function handleMouseDown (ev) {
+		if (isScrollButtonFocused()) {
+			ev.preventDefault();
+		}
+
+		if (props['data-spotlight-container-disabled']) {
+			ev.preventDefault();
+		} else if (type === 'Native') {
+			childAdapter.current.setContainerDisabled(false);
+		}
+	}
+
+// Move to useEventTouch
+
+	function handleTouchStart () {
+		const focusedItem = Spotlight.getCurrent();
+
+		if (!Spotlight.isPaused() && focusedItem && !isScrollButtonFocused()) {
+			focusedItem.blur();
+		}
+	}
+
+// Move to useEventWheel
+
+	function handleWheel ({delta}) {
+		const focusedItem = Spotlight.getCurrent();
+
+		if (focusedItem && !isScrollButtonFocused()) {
+			focusedItem.blur();
+		}
+
+		if (delta !== 0) {
+			mutableRef.current.isWheeling = true;
+			if (!props['data-spotlight-container-disabled']) {
+				childAdapter.current.setContainerDisabled(true);
+			}
+		}
+	}
+
+// Move to Scrollable below
+
+	function isContent (element) {
+		return (element && utilDOM.containsDangerously(uiChildContainerRef, element));
+	}
+
+// Move to useEventFocus
+
+	const startScrollOnFocus = (pos) => {
+		if (pos) {
+			const
+				{top, left} = pos,
+				bounds = uiScrollAdapter.current.getScrollBounds();
+
+		if (type === 'JS') {
+			const
+				scrollHorizontally = bounds.maxLeft > 0 && left !== uiScrollAdapter.current.scrollLeft,
+				scrollVertically = bounds.maxTop > 0 && top !== uiScrollAdapter.current.scrollTop;
+
+			if (scrollHorizontally || scrollVertically) {
+				uiScrollAdapter.current.start({
+					targetX: left,
+					targetY: top,
+					animate: (animationDuration > 0) && spottable.current.animateOnFocus,
+					overscrollEffect: props.overscrollEffectOn[uiScrollAdapter.current.lastInputType] &&
+						(!childAdapter.current.shouldPreventOverscrollEffect || !childAdapter.current.shouldPreventOverscrollEffect())
+				});
+				spottable.current.lastScrollPositionOnFocus = pos;
+			}
+		} else {
+			const
+				scrollHorizontally = bounds.maxLeft > 0 && Math.abs(left - uiScrollAdapter.current.scrollLeft) > epsilon,
+				scrollVertically = bounds.maxTop > 0 && Math.abs(top - uiScrollAdapter.current.scrollTop) > epsilon;
+
+			if (scrollHorizontally || scrollVertically) {
+				uiScrollAdapter.current.start({
+					targetX: left,
+					targetY: top,
+					animate: spottable.current.animateOnFocus,
+					overscrollEffect: props.overscrollEffectOn[uiScrollAdapter.current.lastInputType] &&
+						(!childAdapter.current.shouldPreventOverscrollEffect || !childAdapter.current.shouldPreventOverscrollEffect())
+				});
+				spottable.current.lastScrollPositionOnFocus = pos;
+			}
+		}
+		}
+	};
+
+// Move to useEventFocus
+
+	function calculateAndScrollTo () {
+		const
+			positionFn = childAdapter.current.calculatePositionOnFocus,
+			childContainerNode = uiChildContainerRef.current,
+			spotItem = Spotlight.getCurrent();
+
+		if (spotItem && positionFn && utilDOM.containsDangerously(childContainerNode, spotItem)) {
+			const lastPos = spottable.current.lastScrollPositionOnFocus;
+			let pos;
+
+			// If scroll animation is ongoing, we need to pass last target position to
+			// determine correct scroll position.
+			if (lastPos & (
+				type === 'JS' && uiScrollAdapter.current.animator.isAnimating() ||
+				type === 'Native' && uiScrollAdapter.current.scrolling
+			)) {
+				const
+					containerRect = getRect(childContainerNode),
+					itemRect = getRect(spotItem);
+				let scrollPosition;
+
+				if (props.direction === 'horizontal' || props.direction === 'both' && !(itemRect.left >= containerRect.left && itemRect.right <= containerRect.right)) {
+					scrollPosition = lastPos.left;
+				} else if (props.direction === 'vertical' || props.direction === 'both' && !(itemRect.top >= containerRect.top && itemRect.bottom <= containerRect.bottom)) {
+					scrollPosition = lastPos.top;
+				}
+
+				pos = positionFn({item: spotItem, scrollPosition});
+			} else {
+				// scrollInfo passes in current `scrollHeight` and `scrollTop` before calculations
+				const
+					scrollInfo = {
+						previousScrollHeight: uiScrollAdapter.current.bounds.scrollHeight,
+						scrollTop: uiScrollAdapter.current.scrollTop
+					};
+
+				pos = positionFn({item: spotItem, scrollInfo});
+			}
+
+			if (pos && (pos.left !== uiScrollAdapter.current.scrollLeft || pos.top !== uiScrollAdapter.current.scrollTop)) {
+				startScrollOnFocus(pos);
+			}
+
+			// update `scrollHeight`
+			uiScrollAdapter.current.bounds.scrollHeight = uiScrollAdapter.current.getScrollBounds().scrollHeight;
+		}
+	}
+
+// Move to useEventFocus
+
+	function handleFocus (ev) {
+		const shouldPreventScrollByFocus = childAdapter.current.shouldPreventScrollByFocus ?
+			childAdapter.current.shouldPreventScrollByFocus() :
+			false;
+
+		if (type === 'JS' && isWheeling) {
+			uiScrollAdapter.current.stop();
+			spottable.current.animateOnFocus = false;
+		}
+
+		if (!Spotlight.getPointerMode()) {
+			alertThumb();
+		}
+
+		if (!(shouldPreventScrollByFocus || Spotlight.getPointerMode() || uiScrollAdapter.current.isDragging)) {
+			const
+				item = ev.target,
+				spotItem = Spotlight.getCurrent();
+
+			if (item && item === spotItem) {
+				calculateAndScrollTo();
+			}
+		} else if (childAdapter.current.setLastFocusedNode) {
+			childAdapter.current.setLastFocusedNode(ev.target);
+		}
+	}
+
+// Move to useEventKey
+
+	function scrollByPage (direction) {
+		const
+			{scrollTop} = uiScrollAdapter.current,
+			focusedItem = Spotlight.getCurrent(),
+			bounds = uiScrollAdapter.current.getScrollBounds(),
+			isUp = direction === 'up',
+			directionFactor = isUp ? -1 : 1,
+			pageDistance = directionFactor * bounds.clientHeight * paginationPageMultiplier;
+		let scrollPossible = false;
+
+		if (type === 'JS') {
+			scrollPossible = isUp ? scrollTop > 0 : bounds.maxTop > scrollTop;
+		} else {
+			scrollPossible = isUp ? scrollTop > 0 : bounds.maxTop - scrollTop > epsilon;
+		}
+
+		uiScrollAdapter.current.lastInputType = 'pageKey';
+
+		if (directionFactor !== uiScrollAdapter.current.wheelDirection) {
+			uiScrollAdapter.current.isScrollAnimationTargetAccumulated = false;
+			uiScrollAdapter.current.wheelDirection = directionFactor;
+		}
+
+		if (scrollPossible) {
+			if (focusedItem) {
+				const contentNode = uiChildContainerRef.current;
+
+				// Should do nothing when focusedItem is paging control button of Scrollbar
+				if (utilDOM.containsDangerously(contentNode, focusedItem)) {
+					const
+						contentRect = contentNode.getBoundingClientRect(),
+						clientRect = focusedItem.getBoundingClientRect(),
+						yAdjust = isUp ? 1 : -1,
+						x = clamp(contentRect.left, contentRect.right, (clientRect.right + clientRect.left) / 2);
+					let y = 0;
+
+					if (type === 'JS') {
+						y = bounds.maxTop <= scrollTop + pageDistance || 0 >= scrollTop + pageDistance ?
+							contentRect[isUp ? 'top' : 'bottom'] + yAdjust :
+							clamp(contentRect.top, contentRect.bottom, (clientRect.bottom + clientRect.top) / 2);
+					} else {
+						y = bounds.maxTop - epsilon < scrollTop + pageDistance || epsilon > scrollTop + pageDistance ?
+							contentNode.getBoundingClientRect()[isUp ? 'top' : 'bottom'] + yAdjust :
+							clamp(contentRect.top, contentRect.bottom, (clientRect.bottom + clientRect.top) / 2);
+					}
+
+					focusedItem.blur();
+
+					if (!props['data-spotlight-container-disabled']) {
+						childAdapter.current.setContainerDisabled(true);
+					}
+
+					spottable.current.pointToFocus = {direction, x, y};
+				}
+			} else {
+				spottable.current.pointToFocus = {direction, x: lastPointer.x, y: lastPointer.y};
+			}
+
+			uiScrollAdapter.current.scrollToAccumulatedTarget(pageDistance, true, props.overscrollEffectOn.pageKey);
+		}
+	}
+
+// Move to useEventFocus
+
+	function hasFocus () {
+		let current = Spotlight.getCurrent();
+
+		if (!current) {
+			const spotlightId = Spotlight.getActiveContainer();
+			current = document.querySelector(`[data-spotlight-id="${spotlightId}"]`);
+		}
+
+		return utilDOM.containsDangerously(uiScrollContainerRef, current);
+	}
+
+// Move to useOverscrollEffect
+
+	function checkAndApplyOverscrollEffectByDirection (direction) {
+		const
+			orientation = (direction === 'up' || direction === 'down') ? 'vertical' : 'horizontal',
+			bounds = uiScrollAdapter.current.getScrollBounds(),
+			scrollability = orientation === 'vertical' ? uiScrollAdapter.current.canScrollVertically(bounds) : uiScrollAdapter.current.canScrollHorizontally(bounds);
+
+		if (scrollability) {
+			const
+				isRtl = uiScrollAdapter.current.rtl,
+				edge = (direction === 'up' || !isRtl && direction === 'left' || isRtl && direction === 'right') ? 'before' : 'after';
+
+			uiScrollAdapter.current.checkAndApplyOverscrollEffect(orientation, edge, overscrollTypeOnce);
+		}
+	}
+
+// Move to useEventKey
+
+	function scrollByPageOnPointerMode (ev) {
+		const {keyCode, repeat} = ev;
+
+		forward('onKeyDown', ev, props);
+		ev.preventDefault();
+
+		spottable.current.animateOnFocus = true;
+
+		if (!repeat && (props.direction === 'vertical' || props.direction === 'both')) {
+			const direction = isPageUp(keyCode) ? 'up' : 'down';
+
+			scrollByPage(direction);
+
+			if (props.overscrollEffectOn.pageKey) {
+				checkAndApplyOverscrollEffectByDirection(direction);
+			}
+
+			return true; // means consumed
+		}
+
+		return false; // means to be propagated
+	}
+
+// Move to useEventKey
+
+	function handleKeyDown (ev) {
+		const {keyCode, repeat, target} = ev;
+
+		forward('onKeyDown', ev, props);
+
+		if (isPageUp(keyCode) || isPageDown(keyCode)) {
+			ev.preventDefault();
+		}
+
+		spottable.current.animateOnFocus = true;
+
+		if (!repeat && hasFocus()) {
+			let direction = null;
+
+			if (isPageUp(keyCode) || isPageDown(keyCode)) {
+				if (props.direction === 'vertical' || props.direction === 'both') {
+					direction = isPageUp(keyCode) ? 'up' : 'down';
+
+					if (isContent(target)) {
+						ev.stopPropagation();
+						scrollByPage(direction);
+					}
+					if (props.overscrollEffectOn.pageKey) {
+						checkAndApplyOverscrollEffectByDirection(direction);
+					}
+				}
+			} else if (getDirection(keyCode) && (type === 'JS' || type === 'Native' && !Spotlight.getPointerMode())) {
+				const element = Spotlight.getCurrent();
+
+				uiScrollAdapter.current.lastInputType = 'arrowKey';
+				direction = getDirection(keyCode);
+
+				if (props.overscrollEffectOn.arrowKey && !(element ? getTargetByDirectionFromElement(direction, element) : null)) {
+					if (
+						!(horizontalScrollbarRef.current && utilDOM.containsDangerously(horizontalScrollbarRef.current.uiScrollbarContainer, element)) &&
+						!(verticalScrollbarRef.current && utilDOM.containsDangerously(verticalScrollbarRef.current.uiScrollbarContainer, element))
+					) {
+						checkAndApplyOverscrollEffectByDirection(direction);
+					}
+				}
+			}
+		}
+	}
+
+// Move to useScrollbar
+
+	function onScrollbarButtonClick ({isPreviousScrollButton, isVerticalScrollBar}) {
+		const
+			{wheelDirection} = uiScrollAdapter.current,
+			bounds = uiScrollAdapter.current.getScrollBounds(),
+			direction = isPreviousScrollButton ? -1 : 1,
+			pageDistance = direction * (isVerticalScrollBar ? bounds.clientHeight : bounds.clientWidth) * paginationPageMultiplier;
+
+		uiScrollAdapter.current.lastInputType = 'scrollbarButton';
+
+		if (direction !== wheelDirection) {
+			uiScrollAdapter.current.isScrollAnimationTargetAccumulated = false;
+			uiScrollAdapter.current.wheelDirection = direction;
+		}
+
+		uiScrollAdapter.current.scrollToAccumulatedTarget(pageDistance, isVerticalScrollBar, props.overscrollEffectOn.scrollbarButton);
+	}
+
+	function focusOnScrollButton (scrollbarRef, isPreviousScrollButton) {
+		if (scrollbarRef.current) {
+			scrollbarRef.current.focusOnButton(isPreviousScrollButton);
+		}
+	}
+
+	function scrollAndFocusScrollbarButton (direction) {
+		if (uiScrollAdapter.current) {
+			const
+				{rtl} = uiScrollAdapter.current,
+				isPreviousScrollButton = direction === 'up' || (rtl ? direction === 'right' : direction === 'left'),
+				isHorizontalDirection = direction === 'left' || direction === 'right',
+				isVerticalDirection = direction === 'up' || direction === 'down',
+				canScrollHorizontally = isHorizontalDirection && (props.direction === 'horizontal' || props.direction === 'both'),
+				canScrollingVertically = isVerticalDirection && (props.direction === 'vertical' || props.direction === 'both');
+
+			if (canScrollHorizontally || canScrollingVertically) {
+				onScrollbarButtonClick({
+					isPreviousScrollButton,
+					isVerticalScrollBar: canScrollingVertically
+				});
+
+				if (props.focusableScrollbar) {
+					focusOnScrollButton(
+						canScrollingVertically ? verticalScrollbarRef : horizontalScrollbarRef,
+						isPreviousScrollButton
+					);
+				}
+			}
+		}
+	}
+
+// Move to Scrollable below
+
+	function stop () {
+		if (!props['data-spotlight-container-disabled']) {
+			childAdapter.current.setContainerDisabled(false);
+		}
+
+		focusOnItem();
+		mutableRef.current.lastScrollPositionOnFocus = null;
+		mutableRef.current.isWheeling = false;
+		stopVoice();
+	}
+
+	function focusOnItem () {
+		if (mutableRef.current.indexToFocus !== null && typeof childAdapter.current.focusByIndex === 'function') {
+			childAdapter.current.focusByIndex(mutableRef.current.indexToFocus);
+			mutableRef.current.indexToFocus = null;
+		}
+
+		if (mutableRef.current.nodeToFocus !== null && typeof childAdapter.current.focusOnNode === 'function') {
+			childAdapter.current.focusOnNode(mutableRef.current.nodeToFocus);
+			mutableRef.current.nodeToFocus = null;
+		}
+
+		if (mutableRef.current.pointToFocus !== null) {
+			// no need to focus on pointer mode
+			if (!Spotlight.getPointerMode()) {
+				const
+					{direction, x, y} = mutableRef.current.pointToFocus,
+					position = {x, y},
+					elemFromPoint = document.elementFromPoint(x, y),
+					target =
+						elemFromPoint && elemFromPoint.closest && getIntersectingElement(elemFromPoint.closest(`.${spottableClass}`), uiScrollContainerRef.current) ||
+						getTargetInViewByDirectionFromPosition(direction, position, uiScrollContainerRef.current) ||
+						getTargetInViewByDirectionFromPosition(reverseDirections[direction], position, uiScrollContainerRef.current);
+
+				if (target) {
+					Spotlight.focus(target);
+				}
+			}
+
+			mutableRef.current.pointToFocus = null;
+		}
+	}
+
+	function scrollTo (opt) {
+		mutableRef.current.indexToFocus = (opt.focus && typeof opt.index === 'number') ? opt.index : null;
+		mutableRef.current.nodeToFocus = (opt.focus && opt.node instanceof Object && opt.node.nodeType === 1) ? opt.node : null;
+	}
+
+// Move to useScrollbar
+
+	function alertThumb () {
+		const bounds = uiScrollAdapter.current.getScrollBounds();
+
+		uiScrollAdapter.current.showThumb(bounds);
+		uiScrollAdapter.current.startHidingThumb();
+	}
+
+	function alertThumbAfterRendered () {
+		const spotItem = Spotlight.getCurrent();
+
+		if (!Spotlight.getPointerMode() && isContent(spotItem) && uiScrollAdapter.current.isUpdatedScrollThumb) {
+			alertThumb();
+		}
+	}
+
+// Move to Scrollable below
+
+	function handleResizeWindow () {
+		const focusedItem = Spotlight.getCurrent();
+
+		if (focusedItem) {
+			focusedItem.blur();
+		}
+	}
+
+	function handleScrollerUpdate () {
+		if (uiScrollAdapter.current.scrollToInfo === null) {
+			const scrollHeight = uiScrollAdapter.current.getScrollBounds().scrollHeight;
+
+			if (scrollHeight !== uiScrollAdapter.current.bounds.scrollHeight) {
+				calculateAndScrollTo();
+			}
+		}
+
+		// oddly, Scroller manages uiScrollAdapter.current.bounds so if we don't update it here (it is also
+		// updated in calculateAndScrollTo but we might not have made it to that point), it will be
+		// out of date when we land back in this method next time.
+		uiScrollAdapter.current.bounds.scrollHeight = uiScrollAdapter.current.getScrollBounds().scrollHeight;
+	}
+
+// Move to useOverscrollEffect
+
+	function clearOverscrollEffect (orientation, edge) {
+		mutableRef.current.overscrollJobs[orientation][edge].startAfter(overscrollTimeout, orientation, edge, overscrollTypeNone, 0);
+		uiScrollAdapter.current.setOverscrollStatus(orientation, edge, overscrollTypeNone, 0);
+	}
+
+	const applyOverscrollEffect = useCallback((orientation, edge, type, ratio) => {
+		const nodeRef = overscrollRefs[orientation].current;
+
+		if (nodeRef) {
+			nodeRef.style.setProperty(overscrollRatioPrefix + orientation + edge, ratio);
+
+			if (type === overscrollTypeOnce) {
+				mutableRef.current.overscrollJobs[orientation][edge].start(orientation, edge, overscrollTypeDone, 0);
+			}
+		}
+	}, [overscrollRefs]);
+
+	function createOverscrollJob (orientation, edge) {
+		if (!mutableRef.current.overscrollJobs[orientation][edge]) {
+			mutableRef.current.overscrollJobs[orientation][edge] = new Job(applyOverscrollEffect, overscrollTimeout);
+		}
+	}
+
+	function stopOverscrollJob (orientation, edge) {
+		const job = mutableRef.current.overscrollJobs[orientation][edge];
+
+		if (job) {
+			job.stop();
+		}
+	}
+
+// Move to Scrollable below
+
+	// FIXME setting event handlers directly to work on the V8 snapshot.
+	function addEventListeners (ref) { // `ref` is always `uiChildContainerRef`.
+		utilEvent('focusin').addEventListener(ref, handleFocus);
+
+		if (ref.current) {
+			addVoiceEventListener(ref);
+		}
+	}
+
+	// FIXME setting event handlers directly to work on the V8 snapshot.
+	function removeEventListeners (ref) { // `ref` is always `uiChildContainerRef`.
+		utilEvent('focusin').removeEventListener(ref, handleFocus);
+
+		if (ref.current) {
+			removeVoiceEventListener(ref);
+		}
+	}
+
+// Move to useEventVoice
+
+	const updateFocusAfterVoiceControl = () => {
+		const
+			spotItem = Spotlight.getCurrent(),
+			scrollContainerNode = uiScrollContainerRef.current;
+
+		if (utilDOM.containsDangerously(scrollContainerNode, spotItem)) {
+			const
+				viewportBounds = scrollContainerNode.getBoundingClientRect(),
+				spotItemBounds = spotItem.getBoundingClientRect(),
+				nodes = Spotlight.getSpottableDescendants(scrollContainerNode.dataset.spotlightId),
+				first = mutableRef.current.voiceControlDirection === 'vertical' ? 'top' : 'left',
+				last = mutableRef.current.voiceControlDirection === 'vertical' ? 'bottom' : 'right';
+
+			if (spotItemBounds[last] < viewportBounds[first] || spotItemBounds[first] > viewportBounds[last]) {
+				for (let i = 0; i < nodes.length; i++) {
+					const nodeBounds = nodes[i].getBoundingClientRect();
+
+					if (nodeBounds[first] > viewportBounds[first] && nodeBounds[last] < viewportBounds[last]) {
+						Spotlight.focus(nodes[i]);
+						break;
+					}
+				}
+			}
+		}
+	};
+
+	const isReachedEdge = (scrollPos, ltrBound, rtlBound, isRtl = false) => {
+		const bound = isRtl ? rtlBound : ltrBound;
+		return (bound === 0 && scrollPos === 0) || (bound > 0 && scrollPos >= bound - 1);
+	};
+
+
+	const handleVoice = (e) => {
+		const
+			isHorizontal = (props.direction === 'horizontal'),
+			isRtl = uiScrollAdapter.current.rtl,
+			{scrollTop, scrollLeft} = uiScrollAdapter.current,
+			{maxLeft, maxTop} = uiScrollAdapter.current.getScrollBounds(),
+			verticalDirection = ['up', 'down', 'top', 'bottom'],
+			horizontalDirection = isRtl ? ['right', 'left', 'rightmost', 'leftmost'] : ['left', 'right', 'leftmost', 'rightmost'],
+			movement = ['previous', 'next', 'first', 'last'];
+		let
+			scroll = e && e.detail && e.detail.scroll,
+			index = movement.indexOf(scroll);
+
+		if (index > -1) {
+			scroll = isHorizontal ? horizontalDirection[index] : verticalDirection[index];
+		}
+
+		mutableRef.current.voiceControlDirection = verticalDirection.includes(scroll) && 'vertical' || horizontalDirection.includes(scroll) && 'horizontal' || null;
+
+		// Case 1. Invalid direction
+		if (mutableRef.current.voiceControlDirection === null) {
+			mutableRef.current.isVoiceControl = false;
+		// Case 2. Cannot scroll
+		} else if (
+			(['up', 'top'].includes(scroll) && isReachedEdge(scrollTop, 0)) ||
+			(['down', 'bottom'].includes(scroll) && isReachedEdge(scrollTop, maxTop)) ||
+			(['left', 'leftmost'].includes(scroll) && isReachedEdge(scrollLeft, 0, maxLeft, isRtl)) ||
+			(['right', 'rightmost'].includes(scroll) && isReachedEdge(scrollLeft, maxLeft, 0, isRtl))
+		) {
+			if (window.webOSVoiceReportActionResult) {
+				window.webOSVoiceReportActionResult({voiceUi: {exception: 'alreadyCompleted'}});
+				e.preventDefault();
+			}
+		// Case 3. Can scroll
+		} else {
+			mutableRef.current.isVoiceControl = true;
+
+			if (['up', 'down', 'left', 'right'].includes(scroll)) {
+				const isPreviousScrollButton = (scroll === 'up') || (scroll === 'left' && !isRtl) || (scroll === 'right' && isRtl);
+				onScrollbarButtonClick({isPreviousScrollButton, isVerticalScrollBar: verticalDirection.includes(scroll)});
+			} else { // ['top', 'bottom', 'leftmost', 'rightmost'].includes(scroll)
+				uiScrollAdapter.current.scrollTo({align: verticalDirection.includes(scroll) && scroll || (scroll === 'leftmost' && isRtl || scroll === 'rightmost' && !isRtl) && 'right' || 'left'});
+			}
+
+			e.preventDefault();
+		}
+	};
+
+// Move to Scrollable below
+
+	function handleScroll (ev) {
+		const
+			{scrollLeft: x, scrollTop: y} = ev,
+			{id} = props;
+
+		forward('onScroll', ev, props);
+
+		if (id && contextSharedState && contextSharedState.set) {
+			contextSharedState.set(ev, props);
+			contextSharedState.set(`${id}.scrollPosition`, {x, y});
+		}
+	}
+*/
+
+const useSpottableScroll = (props, instances, context) => {
+	const {childAdapter, uiChildContainerRef, uiScrollAdapter, uiScrollContainerRef} = instances;
+	const {type} = context;
+	const contextSharedState = useContext(SharedState);
+
+	// Mutable value
+
+	const mutableRef = useRef({
+		animateOnFocus: false,
+		indexToFocus: null,
+		lastScrollPositionOnFocus: null,
+		nodeToFocus: null,
+		pointToFocus: null
+	});
+
+	// Hooks
+
+	const {
+		alertThumb,
+		isScrollButtonFocused,
+		onScrollbarButtonClick,
+		scrollAndFocusScrollbarButton,
+		scrollbarProps
+	} = useScrollbar(props, instances, {isContent});
+
+	useSpotlightConfig(props);
+
+	useSpotlightRestore(props, instances);
+
+	const {
+		applyOverscrollEffect,
+		checkAndApplyOverscrollEffectByDirection,
+		clearOverscrollEffect
+	} = useOverscrollEffect({}, instances);
+
+	const {handleWheel, isWheeling} = useEventWheel(props, instances, {isScrollButtonFocused, type});
+
+	const {calculateAndScrollTo, handleFocus, hasFocus} = useEventFocus(props, {...instances, spottable: mutableRef}, {alertThumb, isWheeling, type});
+
+	const {handleKeyDown, lastPointer, scrollByPageOnPointerMode} = useEventKey(props, {...instances, spottable: mutableRef}, {checkAndApplyOverscrollEffectByDirection, hasFocus, isContent, type});
+
+	useEventMonitor({}, instances, {lastPointer, scrollByPageOnPointerMode});
+
+	const {handleFlick, handleMouseDown} = useEventMouse({}, instances, {isScrollButtonFocused, type});
+
+	const {handleTouchStart} = useEventTouch({}, instances, {isScrollButtonFocused});
+
+	const {
+		addVoiceEventListener,
+		removeVoiceEventListener,
+		stopVoice
+	} = useEventVoice(props, instances, {onScrollbarButtonClick});
+
+	// Functions
+
+	function isContent (element) {
+		return (element && utilDOM.containsDangerously(uiChildContainerRef, element));
+	}
+
+	function scrollTo (opt) {
+		mutableRef.current.indexToFocus = (opt.focus && typeof opt.index === 'number') ? opt.index : null;
+		mutableRef.current.nodeToFocus = (opt.focus && opt.node instanceof Object && opt.node.nodeType === 1) ? opt.node : null;
+	}
+
+	function start (animate) {
+		if (type === 'Native' && !animate) {
+			focusOnItem();
+		}
+	}
+
+	function stop () {
+		if (!props['data-spotlight-container-disabled']) {
+			childAdapter.current.setContainerDisabled(false);
+		}
+
+		focusOnItem();
+		mutableRef.current.lastScrollPositionOnFocus = null;
+		mutableRef.current.isWheeling = false;
+		stopVoice();
+	}
+
+	function scrollStopOnScroll () {
+		stop();
+	}
+
+	function focusOnItem () {
+		if (mutableRef.current.indexToFocus !== null && typeof childAdapter.current.focusByIndex === 'function') {
+			childAdapter.current.focusByIndex(mutableRef.current.indexToFocus);
+			mutableRef.current.indexToFocus = null;
+		}
+
+		if (mutableRef.current.nodeToFocus !== null && typeof childAdapter.current.focusOnNode === 'function') {
+			childAdapter.current.focusOnNode(mutableRef.current.nodeToFocus);
+			mutableRef.current.nodeToFocus = null;
+		}
+
+		if (mutableRef.current.pointToFocus !== null) {
+			// no need to focus on pointer mode
+			if (!Spotlight.getPointerMode()) {
+				const
+					{direction, x, y} = mutableRef.current.pointToFocus,
+					position = {x, y},
+					elemFromPoint = document.elementFromPoint(x, y),
+					target =
+						elemFromPoint && elemFromPoint.closest && getIntersectingElement(elemFromPoint.closest(`.${spottableClass}`), uiScrollContainerRef.current) ||
+						getTargetInViewByDirectionFromPosition(direction, position, uiScrollContainerRef.current) ||
+						getTargetInViewByDirectionFromPosition(reverseDirections[direction], position, uiScrollContainerRef.current);
+
+				if (target) {
+					Spotlight.focus(target);
+				}
+			}
+
+			mutableRef.current.pointToFocus = null;
+		}
+	}
+
+	function handleScroll (ev) {
+		const
+			{scrollLeft: x, scrollTop: y} = ev,
+			{id} = props;
+
+		forward('onScroll', ev, props);
+
+		if (id && contextSharedState && contextSharedState.set) {
+			contextSharedState.set(ev, props);
+			contextSharedState.set(`${id}.scrollPosition`, {x, y});
+		}
+	}
+
+	// Callback for scroller updates; calculate and, if needed, scroll to new position based on focused item.
+	function handleScrollerUpdate () {
+		if (uiScrollAdapter.current.scrollToInfo === null) {
+			const scrollHeight = uiScrollAdapter.current.getScrollBounds().scrollHeight;
+
+			if (scrollHeight !== uiScrollAdapter.current.bounds.scrollHeight) {
+				calculateAndScrollTo();
+			}
+		}
+
+		// oddly, Scroller manages uiScrollAdapter.current.bounds so if we don't update it here (it is also
+		// updated in calculateAndScrollTo but we might not have made it to that point), it will be
+		// out of date when we land back in this method next time.
+		uiScrollAdapter.current.bounds.scrollHeight = uiScrollAdapter.current.getScrollBounds().scrollHeight;
+	}
+
+	function handleResizeWindow () {
+		const focusedItem = Spotlight.getCurrent();
+
+		if (focusedItem) {
+			focusedItem.blur();
+		}
+	}
+
+	// FIXME setting event handlers directly to work on the V8 snapshot.
+	function addEventListeners (ref) { // `ref` is always `uiChildContainerRef`.
+		utilEvent('focusin').addEventListener(ref, handleFocus);
+
+		if (ref.current) {
+			addVoiceEventListener(ref);
+		}
+	}
+
+	// FIXME setting event handlers directly to work on the V8 snapshot.
+	function removeEventListeners (ref) { // `ref` is always `uiChildContainerRef`.
+		utilEvent('focusin').removeEventListener(ref, handleFocus);
+
+		if (ref.current) {
+			removeVoiceEventListener(ref);
+		}
+	}
+
+	// Return
+
+	return {
+		addEventListeners,
+		applyOverscrollEffect,
+		clearOverscrollEffect,
+		handleFlick,
+		handleKeyDown,
+		handleMouseDown,
+		handleResizeWindow,
+		handleScroll,
+		handleScrollerUpdate,
+		handleTouchStart,
+		handleWheel,
+		removeEventListeners,
+		scrollAndFocusScrollbarButton,
+		scrollbarProps,
+		scrollStopOnScroll,
+		scrollTo,
+		start,
+		stop
+	};
+};
+
+const useScroll = (props) => {
+			// TBD: indentation is broken intentionally to help comparing
+			/* eslint-disable indent */
+			const {
+				'data-spotlight-container': spotlightContainer,
+				'data-spotlight-container-disabled': spotlightContainerDisabled,
+				'data-spotlight-id': spotlightId,
+				focusableScrollbar,
+				preventBubblingOnKeyDown,
+				scrollDownAriaLabel,
+				scrollLeftAriaLabel,
+				scrollRightAriaLabel,
+				scrollUpAriaLabel,
+				type,
+				...rest
+			} = props,
+			downButtonAriaLabel = scrollDownAriaLabel == null ? $L('scroll down') : scrollDownAriaLabel,
+			upButtonAriaLabel = scrollUpAriaLabel == null ? $L('scroll up') : scrollUpAriaLabel,
+			rightButtonAriaLabel = scrollRightAriaLabel == null ? $L('scroll right') : scrollRightAriaLabel,
+			leftButtonAriaLabel = scrollLeftAriaLabel == null ? $L('scroll left') : scrollLeftAriaLabel;
+
+	// Mutable value
+
+	const uiScrollContainerRef = useRef();
+	const uiChildContainerRef = useRef();
+
+	const overscrollRefs = {
+		horizontal: React.useRef(),
+		vertical: React.useRef()
+	};
+
+	const horizontalScrollbarRef = useRef();
+	const verticalScrollbarRef = useRef();
+
+	// Adapters
+
+	const [childAdapter, setChildAdapter] = useChildAdapter();
+
+	const uiScrollAdapter = useRef({
+		animator: null,
+		applyOverscrollEffect: null,
+		bounds: null,
+		calculateDistanceByWheel: null,
+		canScrollHorizontally: null,
+		canScrollVertically: null,
+		checkAndApplyOverscrollEffect: null,
+		getScrollBounds: null,
+		isDragging: null,
+		isScrollAnimationTargetAccumulated: null,
+		isUpdatedScrollThumb: null,
+		lastInputType: null,
+		rtl: null,
+		scrollBounds: null,
+		scrollHeight: null,
+		scrolling: null,
+		scrollLeft: null,
+		scrollPos: null,
+		scrollTo: null,
+		scrollToAccumulatedTarget: null,
+		scrollToInfo: null,
+		scrollTop: null,
+		setOverscrollStatus: null,
+		showThumb: null,
+		start: null,
+		startHidingThumb: null,
+		stop: null,
+		wheelDirection: null
+	});
+
+	const setUiScrollAdapter = (adapter) => {
+		uiScrollAdapter.current = adapter;
+	};
+
+	const [uiChildAdapter, setUiChildAdapter] = useUiChildAdapter();
+
+	// Hooks
+
+	const instance = {
+		// Ref
+		uiScrollContainerRef,
+		overscrollRefs,
+		uiChildContainerRef,
+		horizontalScrollbarRef,
+		verticalScrollbarRef,
+
+		// Adapter
+		childAdapter,
+		uiScrollAdapter,
+		uiChildAdapter
+	};
+
+	const
+		decoratedChildProps = {},
+		decorateChildProps = utilDecorateChildProps(decoratedChildProps),
+		scrollProps = {};
+
+	const {
+		addEventListeners,
+		applyOverscrollEffect,
+		clearOverscrollEffect,
+		handleFlick,
+		handleKeyDown,
+		handleMouseDown,
+		handleResizeWindow,
+		handleScroll,
+		handleScrollerUpdate,
+		handleTouchStart,
+		handleWheel,
+		removeEventListeners,
+		scrollAndFocusScrollbarButton,
+		scrollbarProps,
+		scrollStopOnScroll, // Native
+		scrollTo,
+		start, // Native
+		stop // JS
+	} = useSpottableScroll(props, instance, {type});
+
+	// Render
+
+	if (type === 'JS') {
+		scrollProps.stop = stop;
+	} else {
+		scrollProps.scrollStopOnScroll = scrollStopOnScroll;
+		scrollProps.start = start;
+	}
+
+	decorateChildProps('scrollContainerProps', {
+		className: [overscrollCss.scroll],
+		'data-spotlight-container': spotlightContainer,
+		'data-spotlight-container-disabled': spotlightContainerDisabled,
+		'data-spotlight-id': spotlightId,
+		onTouchStart: handleTouchStart
+	});
+
+	decorateChildProps('innerScrollContainerProps', {
+		className: [overscrollCss.overscrollFrame, overscrollCss.vertical]
+	});
+
+	decorateChildProps('childWrapperProps', {
+		className: [overscrollCss.overscrollFrame, overscrollCss.horizontal]
+	});
+
+	decorateChildProps('childProps', {
+		onUpdate: handleScrollerUpdate,
+		scrollAndFocusScrollbarButton,
+		setChildAdapter,
+		spotlightId,
+		uiScrollAdapter
+	});
+
+	decorateChildProps('verticalScrollbarProps', {
+		...scrollbarProps,
+		focusableScrollButtons: focusableScrollbar,
+		nextButtonAriaLabel: downButtonAriaLabel,
+		onKeyDownButton: handleKeyDown,
+		preventBubblingOnKeyDown,
+		previousButtonAriaLabel: upButtonAriaLabel
+	});
+
+	decorateChildProps('horizontalScrollbarProps', {
+		...scrollbarProps,
+		focusableScrollButtons: focusableScrollbar,
+		nextButtonAriaLabel: rightButtonAriaLabel,
+		onKeyDownButton: handleKeyDown,
+		preventBubblingOnKeyDown,
+		previousButtonAriaLabel: leftButtonAriaLabel
+	});
+
+	const {
+		childWrapper,
+		isHorizontalScrollbarVisible,
+		isVerticalScrollbarVisible
+	} = useScrollBase({
+		...rest,
+		...scrollProps,
+		decorateChildProps,
+		noScrollByDrag: !platform.touchscreen,
+		addEventListeners,
+		applyOverscrollEffect,
+		clearOverscrollEffect,
+		handleResizeWindow,
+		horizontalScrollbarRef,
+		onFlick: handleFlick,
+		onKeyDown: handleKeyDown,
+		onMouseDown: handleMouseDown,
+		onScroll: handleScroll,
+		onWheel: handleWheel,
+		removeEventListeners,
+		scrollTo,
+		setUiChildAdapter,
+		setUiScrollAdapter,
+		type,
+		uiChildAdapter,
+		uiChildContainerRef,
+		uiScrollContainerRef,
+		verticalScrollbarRef
+	});
+
+	decorateChildProps('innerScrollContainerProps', {
+		className: [...(isHorizontalScrollbarVisible ? overscrollCss.horizontalScrollbarVisible : [])]
+	});
+
+	decorateChildProps('scrollContainerProps', {ref: uiScrollContainerRef});
+	decorateChildProps('innerScrollContainerProps', {ref: overscrollRefs.vertical});
+	decorateChildProps('childWrapperProps', {ref: overscrollRefs.horizontal});
+	decorateChildProps('childProps', {uiChildAdapter, uiChildContainerRef});
+	decorateChildProps('verticalScrollbarProps', {ref: verticalScrollbarRef});
+	decorateChildProps('horizontalScrollbarProps', {ref: horizontalScrollbarRef});
+
+	return {
+		...decoratedChildProps,
+		childWrapper,
+		isHorizontalScrollbarVisible,
+		isVerticalScrollbarVisible
+	};
 };
 
 /**
@@ -188,11 +1332,11 @@ ScrollableBase.defaultProps = {
  * @ui
  * @public
  */
-const Scrollable = {};
 
-export default Scrollable;
+export default useScroll;
 export {
 	dataIndexAttribute,
 	ScrollableBase as Scrollable,
-	ScrollableBase
+	ScrollableBase,
+	useScroll
 };
