+++ packages/moonstone/Scrollable/useScrollbar.js
@@ -3,8 +3,7 @@ import Spotlight from '@enact/spotlight';
 const paginationPageMultiplier = 0.66;
 
 const useScrollbar = (props, instances, context) => {
-	const {direction: directionProp, focusableScrollbar, overscrollEffectOn} = props;
-	const {horizontalScrollbarRef, uiScrollableAdapter, verticalScrollbarRef} = instances;
+	const {horizontalScrollbarRef, uiScrollAdapter, verticalScrollbarRef} = instances;
 	const {isContent} = context;
 
 	const scrollbarProps = {
@@ -24,19 +23,19 @@ const useScrollbar = (props, instances, context) => {
 
 	function onScrollbarButtonClick ({isPreviousScrollButton, isVerticalScrollBar}) {
 		const
-			{wheelDirection} = uiScrollableAdapter.current,
-			bounds = uiScrollableAdapter.current.getScrollBounds(),
+			{wheelDirection} = uiScrollAdapter.current,
+			bounds = uiScrollAdapter.current.getScrollBounds(),
 			direction = isPreviousScrollButton ? -1 : 1,
 			pageDistance = direction * (isVerticalScrollBar ? bounds.clientHeight : bounds.clientWidth) * paginationPageMultiplier;
 
-		uiScrollableAdapter.current.lastInputType = 'scrollbarButton';
+		uiScrollAdapter.current.lastInputType = 'scrollbarButton';
 
 		if (direction !== wheelDirection) {
-			uiScrollableAdapter.current.isScrollAnimationTargetAccumulated = false;
-			uiScrollableAdapter.current.wheelDirection = direction;
+			uiScrollAdapter.current.isScrollAnimationTargetAccumulated = false;
+			uiScrollAdapter.current.wheelDirection = direction;
 		}
 
-		uiScrollableAdapter.current.scrollToAccumulatedTarget(pageDistance, isVerticalScrollBar, overscrollEffectOn.scrollbarButton);
+		uiScrollAdapter.current.scrollToAccumulatedTarget(pageDistance, isVerticalScrollBar, props.overscrollEffectOn.scrollbarButton);
 	}
 
 	function focusOnScrollButton (scrollbarRef, isPreviousScrollButton) {
@@ -46,42 +45,42 @@ const useScrollbar = (props, instances, context) => {
 	}
 
 	function scrollAndFocusScrollbarButton (direction) {
-		const
-			{hRef, rtl, vRef} = uiScrollableAdapter.current,
-			isPreviousScrollButton = direction === 'up' || (rtl ? direction === 'right' : direction === 'left'),
-			isHorizontalDirection = direction === 'left' || direction === 'right',
-			isVerticalDirection = direction === 'up' || direction === 'down',
-			canScrollHorizontally = isHorizontalDirection && (directionProp === 'horizontal' || directionProp === 'both'),
-			canScrollingVertically = isVerticalDirection && (directionProp === 'vertical' || directionProp === 'both');
+		if (uiScrollAdapter.current) {
+			const
+				{rtl} = uiScrollAdapter.current,
+				isPreviousScrollButton = direction === 'up' || (rtl ? direction === 'right' : direction === 'left'),
+				isHorizontalDirection = direction === 'left' || direction === 'right',
+				isVerticalDirection = direction === 'up' || direction === 'down',
+				canScrollHorizontally = isHorizontalDirection && (props.direction === 'horizontal' || props.direction === 'both'),
+				canScrollingVertically = isVerticalDirection && (props.direction === 'vertical' || props.direction === 'both');
 
-		if (canScrollHorizontally || canScrollingVertically) {
-			onScrollbarButtonClick({
-				isPreviousScrollButton,
-				isVerticalScrollBar: canScrollingVertically
-			});
+			if (canScrollHorizontally || canScrollingVertically) {
+				onScrollbarButtonClick({
+					isPreviousScrollButton,
+					isVerticalScrollBar: canScrollingVertically
+				});
 
-			if (focusableScrollbar) {
-				focusOnScrollButton(
-					canScrollingVertically ? vRef : hRef,
-					isPreviousScrollButton
-				);
+				if (props.focusableScrollbar) {
+					focusOnScrollButton(
+						canScrollingVertically ? verticalScrollbarRef : horizontalScrollbarRef,
+						isPreviousScrollButton
+					);
+				}
 			}
 		}
 	}
 
 	function alertThumb () {
-		const bounds = uiScrollableAdapter.current.getScrollBounds();
+		const bounds = uiScrollAdapter.current.getScrollBounds();
 
-		uiScrollableAdapter.current.showThumb(bounds);
-		uiScrollableAdapter.current.startHidingThumb();
+		uiScrollAdapter.current.showThumb(bounds);
+		uiScrollAdapter.current.startHidingThumb();
 	}
 
 	function alertThumbAfterRendered () {
-		const
-			{isUpdatedScrollThumb} = uiScrollableAdapter.current,
-			spotItem = Spotlight.getCurrent();
+		const spotItem = Spotlight.getCurrent();
 
-		if (!Spotlight.getPointerMode() && isContent(spotItem) && isUpdatedScrollThumb) {
+		if (!Spotlight.getPointerMode() && isContent(spotItem) && uiScrollAdapter.current.isUpdatedScrollThumb) {
 			alertThumb();
 		}
 	}
