--- packages/moonstone/Scrollable/utilEventResizeWindow.js
@@ -1,24 +1,0 @@
-import Spotlight from '@enact/spotlight';
-
-const utilEventResizeWindow = () => {
-	// Functions
-
-	function handleResizeWindow () {
-		const focusedItem = Spotlight.getCurrent();
-
-		if (focusedItem) {
-			focusedItem.blur();
-		}
-	}
-
-	// Return
-
-	return {
-		handleResizeWindow
-	};
-};
-
-export default utilEventResizeWindow;
-export {
-	utilEventResizeWindow
-};
