+++ packages/moonstone/Scroller/useEvent.js
@@ -4,20 +4,20 @@ import {useRef} from 'react';
 const useEventKey = () => {
 	// Mutable value
 
-	const variables = useRef({
+	const mutableRef = useRef({
 		fn: null
 	});
 
 	// Functions
 
 	function addGlobalKeyDownEventListener (fn) {
-		variables.current.fn = fn;
-		utilEvent('keydown').addEventListener(document, variables.current.fn, {capture: true});
+		mutableRef.current.fn = fn;
+		utilEvent('keydown').addEventListener(document, mutableRef.current.fn, {capture: true});
 	}
 
 	function removeGlobalKeyDownEventListener () {
-		utilEvent('keydown').removeEventListener(document, variables.current.fn, {capture: true});
-		variables.current.fn = null;
+		utilEvent('keydown').removeEventListener(document, mutableRef.current.fn, {capture: true});
+		mutableRef.current.fn = null;
 	}
 
 	// Return
