+++ packages/moonstone/Scroller/useSpotlight.js
@@ -2,8 +2,7 @@ import Spotlight from '@enact/spotlight';
 import {useEffect} from 'react';
 
 const useSpotlightConfig = (props, instances) => {
-	const {dangerouslyContainsInScrollable, scrollAndFocusScrollbarButton, spotlightId} = props;
-	const {uiChildAdapter} = instances;
+	const {uiScrollAdapter} = instances;
 
 	// Hooks
 
@@ -11,32 +10,31 @@ const useSpotlightConfig = (props, instances) => {
 		function handleLeaveContainer ({direction, target}) {
 			// ensure we only scroll to boundary from the contents and not a scroll button which
 			// lie outside of uiChildContainerRef but within the spotlight container
-			if (dangerouslyContainsInScrollable(target)) {
+			if (props.scrollContainerContainsDangerously(target)) {
 				const
-					{scrollBounds: {maxLeft, maxTop}, scrollPos: {left, top}} = uiChildAdapter.current,
+					{scrollBounds: {maxLeft, maxTop}, scrollPos: {left, top}} = uiScrollAdapter.current,
 					isVerticalDirection = (direction === 'up' || direction === 'down'),
 					pos = isVerticalDirection ? top : left,
 					max = isVerticalDirection ? maxTop : maxLeft;
 
 				// If max is equal to 0, it means scroller can not scroll to the direction.
 				if (pos >= 0 && pos <= max && max !== 0) {
-					scrollAndFocusScrollbarButton(direction);
+					props.scrollAndFocusScrollbarButton(direction);
 				}
 			}
 		}
 
 		function configureSpotlight () {
-			Spotlight.set(spotlightId, {
+			Spotlight.set(props.spotlightId, {
 				onLeaveContainer: handleLeaveContainer,
 				onLeaveContainerFail: handleLeaveContainer
 			});
 		}
 
 		configureSpotlight();
-	}, [dangerouslyContainsInScrollable, scrollAndFocusScrollbarButton, spotlightId, uiChildAdapter]);
+	}, [props, uiScrollAdapter]);
 };
 
-export default useSpotlightConfig;
 export {
 	useSpotlightConfig
 };
