+++ packages/moonstone/styles/colors-light.less
@@ -210,7 +210,7 @@
 
 // Scrollable
 // ---------------------------------------
-@moon-scrollable-overscroll-color: #454545;
+@moon-scroll-overscroll-color: #454545;
 
 // Slider
 // ---------------------------------------
