+++ packages/moonstone/styles/colors.less
@@ -236,7 +236,7 @@
 
 // Scrollable
 // ---------------------------------------
-@moon-scrollable-overscroll-color: @moon-gray;
+@moon-scroll-overscroll-color: @moon-gray;
 
 
 // Scroller
