+++ packages/moonstone/styles/variables.less
@@ -369,9 +369,9 @@
 
 // Scrollable
 // ---------------------------------------
-@moon-scrollable-overscroll-thickness: 6px;
-@moon-scrollable-overscroll-duration: 300ms;
-@moon-scrollable-scrollbar-size: @moon-icon-button-small-size;
+@moon-scroll-overscroll-thickness: 6px;
+@moon-scroll-overscroll-duration: 300ms;
+@moon-scroll-scrollbar-size: @moon-icon-button-small-size;
 
 // ToggleSwitch
 // ---------------------------------------
