+++ packages/ui/Scrollable/Scrollable.module.less
@@ -1,18 +1,18 @@
 /* Scrollable.module.less */
-.scrollableFill {
+.scrollFill {
 	position: relative;
 	width: 100%;
 	height: 100%;
 }
 
-.scrollable {
-	.scrollableFill;
+.scroll {
+	.scrollFill;
 
 	display: flex;
 	box-sizing: border-box;
 	flex-direction: column;
 
-	.flexLayout {
+	.innerScrollContainer {
 		display: flex;
 		width: 100%;
 		height: 100%;
