--- packages/ui/Scrollable/useForceUpdate.js
@@ -1,7 +1,0 @@
-import {useReducer} from 'react';
-
-const useForceUpdate = () => {
-	return useReducer(x => x + 1, 0);
-};
-
-export default useForceUpdate;
